+++ jt-808-server-support/src/test/java/io/github/hylexus/jt/jt808/spec/builtin/msg/BaseReqRespMsgTest.java
@@ -1,6 +1,7 @@
 package io.github.hylexus.jt.jt808.spec.builtin.msg;
 
 import io.github.hylexus.jt.jt808.spec.Jt808MsgBuilder;
+import io.github.hylexus.jt.jt808.spec.Jt808MsgEncryptionHandler;
 import io.github.hylexus.jt.jt808.spec.Jt808Request;
 import io.github.hylexus.jt.jt808.spec.impl.BuiltinJt808MsgTypeParser;
 import io.github.hylexus.jt.jt808.spec.impl.DefaultJt808ProtocolVersionDetector;
@@ -23,7 +24,8 @@ public class BaseReqRespMsgTest {
     protected final Jt808MsgDecoder jt808MsgDecoder = new DefaultJt808MsgDecoder(
             new BuiltinJt808MsgTypeParser(),
             new DefaultJt808MsgBytesProcessor(ByteBufAllocator.DEFAULT),
-            new DefaultJt808ProtocolVersionDetectorRegistry(new DefaultJt808ProtocolVersionDetector())
+            new DefaultJt808ProtocolVersionDetectorRegistry(new DefaultJt808ProtocolVersionDetector()),
+            Jt808MsgEncryptionHandler.NO_OPS
     );
 
     protected final Jt808AnnotationBasedDecoder annotationBasedDecoder = new Jt808AnnotationBasedDecoder(new DefaultJt808FieldDeserializerRegistry(true));