+++ samples/jt-808-server-sample-debug/src/main/java/io/github/hylexus/jt/jt808/samples/debug/Jt808ServerSampleDebugApplication.java
@@ -1,5 +1,6 @@
 package io.github.hylexus.jt.jt808.samples.debug;
 
+import io.netty.util.ResourceLeakDetector;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
@@ -10,6 +11,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 public class Jt808ServerSampleDebugApplication {
 
     public static void main(String[] args) {
+        // FIXME 如果你不了解 ResourceLeakDetector 是做什么的,请务必注释掉下面这行代码!!!
+        ResourceLeakDetector.setLevel(ResourceLeakDetector.Level.PARANOID);
         SpringApplication.run(Jt808ServerSampleDebugApplication.class, args);
     }